[Bug 697] move towards using swSize and swIndex to keep-up with changing MATLAB APIs
git-svn-id: svn+ssh://svn.openmicroscopy.org.uk/home/svn/ome/trunk@7853 61f9eca8-113b-463c-a1a4-bab9b0a29f14

+++ src/matlab/OME/Statistics/HaralickTextures/HaralickTextures.c
@@ -64,8 +64,8 @@ void mexFunction(int nlhs, mxArray* plhs[], int nrhs, const mxArray* prhs[])
 	else
 		img = prhs[0];
 	
-	mrows = mxGetM(img) ;
-	ncols = mxGetN(img) ;
+	mrows = (int) mxGetM(img) ;
+	ncols = (int) mxGetN(img) ;
 	
 	if (!(mrows > 1) || !(ncols > 1))
 		mexErrMsgTxt("mb_texture requires an input image, not a scalar.\n") ;
